+++ apps/minifront/src/components/tx-details/tx-viewer.tsx
